+++ cluster/lib/actor_cluster.dart
@@ -1,4 +1,4 @@
-export 'src/cluster.dart' show ActorCluster, NodeState, AfterClusterInit;
+export 'src/cluster.dart' show ActorCluster, NodeState, InitCluster, AddActorFactory, AddActorFactories;
 export 'src/config.dart' show Config, ConfigNode, readConfigFromYaml;
 export 'src/context.dart' show ClusterContext;
 export 'src/ser_des.dart' show SerDes;